new version of student service

+++ src/main/java/com/example/springbootlearning/studentapi/service/StudentService.java
@@ -3,6 +3,7 @@ package com.example.springbootlearning.studentapi.service;
 import com.example.springbootlearning.studentapi.exception.StudentNotFoundException;
 import com.example.springbootlearning.studentapi.model.Student;
 import com.example.springbootlearning.studentapi.repository.StudentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +12,7 @@ import java.util.Optional;
 @Service
 public class StudentService {
 
+    @Autowired
     private final StudentRepository studentRepository;
 
 
@@ -25,15 +27,45 @@ public class StudentService {
 
 
     public Student getStudentById(Long id) {
-        return studentRepository.findById(id).orElseThrow(() -> new StudentNotFoundException(id));
+        return studentRepository.findById(id).orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + id));
     }
 
     public Student saveStudent(Student student) {
         return (Student) studentRepository.save(student);
     }
 
-    public void deleteStudentById(Long id) {
-        studentRepository.deleteById(id);
+    public boolean deleteStudentById(Long id) {
+        if (studentRepository.existsById(id)) {
+            studentRepository.deleteById(id);
+            return true;
+        }else {
+            return false;
+        }
+    }
+
+    public Student updateStudent(Long id,Student studentDetails) {
+        //check if student exist
+        Student student = getStudentById(id);
+
+
+        //update student information
+
+        student.setName(studentDetails.getName());
+        student.setEmail(studentDetails.getEmail());
+        student.setAge(studentDetails.getAge());
+
+        return studentRepository.save(student);
+    }
+
+    public List<Student> getStudentsByName(String name) {
+        List<Student> students = studentRepository.findByNameContainingIgnoreCase(name);
+
+//        if (students.isEmpty()) {
+//            throw new StudentNotFoundException("Student not found");
+//        }
+
+
+        return students;
     }
 
 }
